Fix IINA 1.2.0 crashes instantly (Big Sur), iina#3478
This commit will change the method `SleepPreventer.preventSleep`
to submit a task to the main DispatchQueue to display an alert if macOS
returns an error when IINA asks it to not put the display to sleep.

This fixes a crash due to trying to display an alert using a thread other than
the main thread.

Logging has been added to log the failure macOS is returning.