Fix missing dependencies Xcode build warning
Xcode 14 added this warning:
Run script build phase 'Copy Dylib Symlinks' will be run during every
build because it does not specify any outputs. To address this warning,
either add output dependencies to the script phase, or configure it to
run in every build by unchecking "Based on dependency analysis" in the
script phase.

It is not possible to specify a static list of dependences for this
script so the fix disables the "Based on dependency analysis" setting
for this build phase.